requirementsface.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
face_recognition==1.3.0
Pillow==10.0.0
numpy==1.24.4
dlib==19.24.0
